+++ Controller/FormWebsiteController.php
@@ -53,15 +53,17 @@ class FormWebsiteController extends DefaultController
         $template = $structure->getKey();
 
         $typeClass = $this->getTypeClass($template);
+
         /** @var AbstractType $type */
-        $type = $this->get('form.registry')->getType($typeClass);
+        $type = $this->get('form.registry')->getType($typeClass)->getInnerType();
         $type->setAttributes($attributes);
 
         $this->form = $this->get('form.factory')->create($typeClass);
+
         $this->form->handleRequest($request);
 
         if ($this->form->isSubmitted()
-            && !$this->form->isValid()
+            && $this->form->isValid()
             && $response = $this->handleFormSubmit($request, $type, $attributes)
         ) {
             // success form submit
@@ -93,6 +95,7 @@ class FormWebsiteController extends DefaultController
         $typeClass = $this->getTypeClass($key);
         /** @var AbstractType $type */
         $type = $this->get('form.registry')->getType($typeClass);
+
         $this->form = $this->get('form.factory')->create($typeClass);
         $this->form->handleRequest($request);