+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.0-RC2
 
+ - ENHANCEMENT #131    Add created date to list view
  - BUGFIX      #130    Fixed csv export for big results
  - FEATURE     #129    Add disabled field type configuration