+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev-master
 
+ - HOTFIX      #135    Fix exception on page when form was deleted
  - HOTFIX      #136    Fix default value for choice form types and placeholder
 
 ## 1.0.0-RC2