+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## dev-master
 
  - BUGFIX      #137    Add minimum requirement for twig
+ - HOTFIX      #135    Fix exception on page when form was deleted
+ - HOTFIX      #136    Fix default value for choice form types and placeholder
 
 ## 1.0.0-RC2