+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev-master
 
+ - BUGFIX      #137    Add minimum requirement for twig
  - HOTFIX      #135    Fix exception on page when form was deleted
  - HOTFIX      #136    Fix default value for choice form types and placeholder