+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.0-RC3
 
+ - ENHANCEMENT #143    Remove unused parameter dynamic default view
  - BUGFIX      #142    Render placeholder only with value and upate ChoiceTrait
  - BUGFIX      #141    Activate save button on block add
  - ENHANCEMENT #140    Disable csrf token protection when recaptcha field is used