+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.0-RC3
 
+ - BUGFIX      #144    Avoid normalize keys for dynamic width and ajax templates configuration
  - BUGFIX      #142    Render placeholder only with value and upate ChoiceTrait
  - BUGFIX      #141    Activate save button on block add
  - ENHANCEMENT #140    Disable csrf token protection when recaptcha field is used