+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+ - HOTFIX      #161    Fix problems when no translation exist for field
  - BUGFIX      #156    Fix static form handling after refractoring
  - BUGFIX      #150    Fix implement dropzone documentation
  - ENHANCEMENT #147    Improve upgrade errors for 0.2.0 version